before changing contracts
