+++ src/constants/parameterizer.ts
@@ -1,12 +1,12 @@
 export enum ParameterDefaults {
   MIN_DEPOSIT = 10,
   P_MIN_DEPOSIT = 100,
-  APPLY_STAGE_LENGTH = 600,
-  P_APPLY_STAGE_LENGTH = 1200,
-  COMMIT_STAGE_LENGTH = 600,
-  P_COMMIT_STAGE_LENGTH = 1200,
-  REVEAL_STAGE_LENGTH = 600,
-  P_REVEAL_STAGE_LENGTH = 1200,
+  APPLY_STAGE_LENGTH = 0,
+  P_APPLY_STAGE_LENGTH = 0,
+  COMMIT_STAGE_LENGTH = 0,
+  P_COMMIT_STAGE_LENGTH = 0,
+  REVEAL_STAGE_LENGTH = 0,
+  P_REVEAL_STAGE_LENGTH = 0,
   DISPENSATION_PCT = 50,
   P_DISPENSATION_PCT = 50,
   VOTE_QUORUM = 50,
